perf(plugin-shikiji): improve `highlight`
