+++ plugins/plugin-shikiji/src/node/highlight/index.ts
@@ -1,0 +1,1 @@
+export * from './highlight.js'
